mssql 参数类型 在排序依据列表中多次指定了某一列。排序依据列表中的列必须是唯一的

+++ data/view/cnf/def.go
@@ -46,7 +46,7 @@ var TypeMysqlDicMp = map[string]string{
 	"mediumblob":          "[]byte",
 	"longblob":            "[]byte",
 	"integer":             "int64",
-	"numeric":             "int64",
+	"numeric":             "float64",
 	"smalldatetime":       "time.Time", //sqlserver
 	"nvarchar":            "string",
 	"real":                "float32",
